+++ app/src/main/java/kite1412/irrigo/model/WaterCapacityConfig.kt
@@ -1,5 +1,5 @@
 package kite1412.irrigo.model
 
 data class WaterCapacityConfig(
-    val minWaterCapacityPercent: Float, // 0 - 100%
+    val minWaterCapacityPercent: Double, // 0 - 100%
 )
